move logic to new schema file
