+++ app/components/ContactForm.tsx
@@ -11,8 +11,7 @@ import {
     Box,
 } from '@mui/material'
 import PhoneInput from './PhoneNumberInput'
-import { z } from 'zod'
-import { PhoneNumberUtil } from 'google-libphonenumber'
+import { ContactUsFormSchema } from '../schemas/contact-us.schema'
 
 interface FormDataProps {
     firstName: string
@@ -24,80 +23,8 @@ interface FormDataProps {
     countryCode: string
 }
 
-const phoneUtil = PhoneNumberUtil.getInstance()
-
 export default function ContactForm() {
     const [formErrors, setFormErrors] = useState<{ [key: string]: string }>({})
-    const characterLimit: number = 100
-    const phoneNumberLimit: number = 15
-
-    const contactUsFormSchema = z
-        .object({
-            firstName: z
-                .string({
-                    invalid_type_error: 'First Name must be a string',
-                })
-                .min(1, 'First Name is required')
-                .max(
-                    characterLimit,
-                    `First Name must be at most ${characterLimit} characters long`,
-                ),
-            lastName: z
-                .string({
-                    invalid_type_error: 'Last Name must be a string',
-                })
-                .min(1, 'Last Name is required')
-                .max(
-                    characterLimit,
-                    `Last Name must be at most ${characterLimit} characters long`,
-                ),
-            company: z
-                .string({
-                    invalid_type_error: 'Company must be a string',
-                })
-                .min(1, 'Company is required')
-                .max(
-                    characterLimit,
-                    `Company must be at most ${characterLimit} characters long`,
-                ),
-            email: z
-                .string({
-                    invalid_type_error: 'email must be a string',
-                })
-                .min(1, 'email is required')
-                .email('email must be a valid email'),
-            phoneNumber: z
-                .string({
-                    invalid_type_error: 'Phone number must be a string',
-                })
-                .min(1, 'Phone number is required')
-                .max(
-                    phoneNumberLimit,
-                    `Phone number must be at most ${phoneNumberLimit} characters long`,
-                )
-                // Custom validation: Phone number must be numeric
-                .refine((phoneNumber) => /^\d+$/.test(phoneNumber), {
-                    message: 'Phone number must contain only digits',
-                }),
-            phonePrefix: z.string().min(1, 'Phone Prefix is required'),
-            countryCode: z.string().min(1, 'Country Code is required'),
-        })
-        .superRefine((data, ctx) => {
-            const { phonePrefix, countryCode, phoneNumber } = data
-
-            const isValid = phoneUtil.isValidNumberForRegion(
-                phoneUtil.parseAndKeepRawInput(phoneNumber, countryCode),
-                countryCode,
-            )
-
-            if (!isValid) {
-                ctx.addIssue({
-                    code: 'custom',
-                    path: ['phoneNumber'],
-                    message: `${phonePrefix}${phoneNumber} is not valid phone number`,
-                })
-            }
-        })
 
     const handleSubmit = (event: React.MouseEvent<HTMLButtonElement>) => {
         event.preventDefault() // Prevent form submission
@@ -112,7 +39,7 @@ export default function ContactForm() {
             formData.entries(),
         ) as unknown as FormDataProps
 
-        const result = contactUsFormSchema.safeParse(values)
+        const result = ContactUsFormSchema.safeParse(values)
 
         if (!result.success) {
             const formattedErrors = result.error.errors.reduce(
